// front-end/src/api/models/api/dailyTodoItem.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Wijckie API
 * Default API for Wijckie
 * OpenAPI spec version: 1.0.0
 */
import type { StatusEnum } from "./statusEnum"

export interface DailyTodoItem {
    readonly id: number
    readonly module: number
    readonly date: string
    readonly option: number
    status: StatusEnum
}