// front-end/src/api/models/api/statusEnum.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Wijckie API
 * Default API for Wijckie
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `todo` - Todo
 * `skip` - Skip
 * `done` - Done
 */
export type StatusEnum = (typeof StatusEnum)[keyof typeof StatusEnum]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StatusEnum = {
    todo: "todo",
    skip: "skip",
    done: "done",
} as const